Part 7 - Guest User access
- Broke StorageLocation into its own file to fix circular dependency (based on imports)
